requirements.txt
antlr4-python3-runtime==4.8
anyio==3.5.0
asgiref==3.4.1
certifi==2020.6.20
chardet==3.0.4
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
Cython==0.29.28
dataclasses==0.8
distlib==0.3.4
fastapi==0.75.1
filelock==3.4.1
fvcore==0.1.5.post20220305
h11==0.13.0
httptools==0.3.0
idna==2.8
immutables==0.17
importlib-metadata==4.8.3
importlib-resources==5.4.0
iopath==0.1.9
kiwisolver==1.3.1
matplotlib==3.3.4
nptyping==1.4.4
numpy==1.19.5
omegaconf==2.1.1
opencv-python==4.5.5.64
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
portalocker==2.4.0
py==1.11.0
pycocotools-windows==2.0.0.2
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pywin32==303
PyYAML==6.0
requests==2.22.0
six==1.16.0
sniffio==1.2.0
-e .
starlette==0.17.1
tabulate==0.8.9
termcolor==1.1.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tox==3.24.5
tqdm==4.64.0
typing_extensions==4.1.1
typish==1.9.3
urllib3==1.25.11
uvicorn==0.16.0
virtualenv==20.14.0
watchgod==0.7
websockets==9.1
wincertstore==0.2
yacs==0.1.8
zipp==3.6.0
pytest